chore: reorganize landing page
